+++ src/TestStartup/StartupExecution_Tests.cs
@@ -1,0 +1,15 @@
+[assembly: KiboardsTestStartup("TestStartup.Startup")]
+[assembly: TestFramework("KiBoards.TestFramework", "KiBoards.Xunit")]
+
+namespace TestStartup
+{
+    public class StartupExecution_Tests
+    {
+
+        [Fact]
+        public void CreateInstance()
+        {
+            Assert.True(Startup.InstanceCreated);
+        }
+    }
+}
